* 'master' of https://bitbucket.org/lei-isep/lapr3-2020-g033:
  AUCMN1-213 #comment Tests done
  AUCMN1-213 #comment tests
  AUCMN1-213 #comment Tests done
  AUCMN1-209 #comment updated some classes
  AUCMN1-209 #comment updated some classes
  Merge branch 'master' of https://bitbucket.org/lei-isep/lapr3-2020-g033
  AUCMN1-213 #comment Tests done
  AUCMN1-213 #comment tests corrected
  AUCMN1-167 #comment Writing the Main function in C and correcting some issues in the Assembly function
  AUCMN1-213 #comment tests class delivery
  AUCMN1-161 #progress on UC15 [development]
  AUCMN1-212 #comment credit card test
  AUCMN1-208 #comment classes updated
  AUCMN1-167 #implementation Updates

# Conflicts:
#	src/test/java/lapr/project/controller/OrderControllerTest.java

+++ src/main/java/lapr/project/controller/UserController.java
@@ -4,6 +4,7 @@ import lapr.project.data.CourierDataHandler;
 import lapr.project.data.UserDataHandler;
 import lapr.project.model.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class UserController {
@@ -20,7 +21,7 @@ public class UserController {
         return user;
     }
 
-    public void addUserAsClient(String name, String email, String pwd, String role, int nif, long creditCardNumber, int creditCardMonthExpiration, int creditCardNumberYearExpiration, int ccv, double latitude, double longitude, String street, int doorNum, String zipcode, String locality) {
+    public void addUserAsClient(String name, String email, String pwd, String role, int nif, BigDecimal creditCardNumber, int creditCardMonthExpiration, int creditCardNumberYearExpiration, int ccv, double latitude, double longitude, String street, int doorNum, String zipcode, String locality) {
         Address add = new Address(latitude, longitude, street, doorNum, zipcode, locality);
         add.save();
         CreditCard credcard = new CreditCard(creditCardNumber, creditCardMonthExpiration, creditCardNumberYearExpiration, ccv);
@@ -31,7 +32,7 @@ public class UserController {
         userAsClient.save();
     }
 
-    public void addUserAsCourier(String name, String email, int nif, double nss, String password, double maxWeightCapacity, double weight, int pharmacyID, String courierRole) {
+    public void addUserAsCourier(String name, String email, int nif, BigDecimal nss, String password, double maxWeightCapacity, double weight, int pharmacyID, String courierRole) {
         User userAsCourier = new User(email, password, courierRole);
         userAsCourier.save();
        Courier courier = new Courier(email, name, nif, nss, maxWeightCapacity, weight, pharmacyID);
