+++ src/test/java/lapr/project/utils/DoPaymentTest.java
@@ -3,7 +3,7 @@ package lapr.project.utils;
 import lapr.project.model.Client;
 import org.junit.jupiter.api.Test;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -17,7 +17,7 @@ class DoPaymentTest {
 
     @Test
     void doesPayment() {
-        boolean result = dp.doesPayment(new Client(1, "Jorge", "jorge@gmail.com", "qwerty", 132456789, 11,1,new BigInteger("1234567891057189").intValue()), 10);
+        boolean result = dp.doesPayment(new Client(1, "Jorge", "jorge@gmail.com", "qwerty", 132456789, 11,1,new BigDecimal("1234567891057189")), 10);
         boolean actualResult = true;
         assertEquals(result, actualResult);
